--- components/ILIAS/Block/maintenance.json
@@ -1,16 +1,0 @@
-{
-    "maintenance_model": "Classic",
-    "first_maintainer": "",
-    "second_maintainer": "",
-    "implicit_maintainers": [
-        ""
-    ],
-    "coordinator": [
-        ""
-    ],
-    "tester": "",
-    "testcase_writer": "",
-    "path": "Services/Block",
-    "belong_to_component": "None",
-    "used_in_components": []
-}
