+++ components/ILIAS/Container/Block/Service/class.InternalDomainService.php
@@ -18,7 +18,7 @@
 
 declare(strict_types=1);
 
-namespace ILIAS\Block;
+namespace ILIAS\Container\Block;
 
 use ILIAS\DI\Container;
 use ILIAS\Repository\GlobalDICDomainServices;
