+++ components/ILIAS/Container/Block/class.BlockSessionRepository.php
@@ -18,7 +18,7 @@
 
 declare(strict_types=1);
 
-namespace ILIAS\Block;
+namespace ILIAS\Container\Block;
 
 /**
  * Stores repository clipboard data
