+++ components/ILIAS/Container/Container.php
@@ -36,5 +36,7 @@ class Container implements Component\Component
             new Component\Resource\ComponentJS($this, "Container.js");
         $contribute[Component\Resource\PublicAsset::class] = fn() =>
             new Component\Resource\ComponentJS($this, "ilClassification.js");
+        $contribute[Component\Resource\PublicAsset::class] = fn() =>
+            new Component\Resource\ComponentJS($this, "ilblockcallback.js");
     }
 }
